--- src/containers/tud/schemas/followup/follow.js
@@ -1,26 +1,0 @@
-// @flow
-
-import { DataProcessingUtils } from 'lattice-fabricate';
-
-const { getPageSectionKey } = DataProcessingUtils;
-
-const createSchema = (pageNum :number) => ({
-  [getPageSectionKey(pageNum, 1)]: {
-    type: 'object',
-    title: '',
-    properties: {
-
-    }
-  }
-});
-
-const createUiSchema = (pageNum :number) => ({
-  [getPageSectionKey(pageNum, 1)]: {
-    classNames: 'column-span-12 grid-container'
-  }
-});
-
-export {
-  createSchema,
-  createUiSchema
-};
